Separate to mutaton and use it
