+++ lib/domain/diary_setting_physical_condtion_detail/mutation.dart
@@ -1,0 +1,42 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:pilll/database/database.dart';
+import 'package:pilll/entity/diary_setting.codegen.dart';
+import 'package:pilll/util/datetime/day.dart';
+
+final createDiarySettingPhysicalConditionDetailProvider =
+    Provider.autoDispose((ref) => CreateDiarySettingPhysicalConditionDetail(ref.watch(databaseProvider).diarySettingReference()));
+
+class CreateDiarySettingPhysicalConditionDetail {
+  final DocumentReference<DiarySetting> reference;
+  CreateDiarySettingPhysicalConditionDetail(this.reference);
+
+  Future<void> call() async {
+    await reference.set(DiarySetting(createdAt: now()), SetOptions(merge: true));
+  }
+}
+
+final addDiarySettingPhysicalConditionDetailProvider =
+    Provider.autoDispose((ref) => AddDiarySettingPhysicalConditionDetail(ref.watch(databaseProvider).diarySettingReference()));
+
+class AddDiarySettingPhysicalConditionDetail {
+  final DocumentReference<DiarySetting> reference;
+  AddDiarySettingPhysicalConditionDetail(this.reference);
+
+  Future<void> call({required DiarySetting diarySetting, required String physicalConditionDetail}) async {
+    await reference.set(
+        diarySetting.copyWith(physicalConditions: diarySetting.physicalConditions..add(physicalConditionDetail)), SetOptions(merge: true));
+  }
+}
+
+final deleteDiarySettingPhysicalConditionDetailProvider = Provider.autoDispose((ref) => ref.watch(databaseProvider).diarySettingReference());
+
+class DeleteDiarySettingPhysicalConditionDetail {
+  final DocumentReference<DiarySetting> reference;
+  DeleteDiarySettingPhysicalConditionDetail(this.reference);
+
+  Future<void> call({required DiarySetting diarySetting, required List<String> physicalConditionDetails}) async {
+    await reference.set(diarySetting.copyWith(physicalConditions: diarySetting.physicalConditions..removeWhere(physicalConditionDetails.contains)),
+        SetOptions(merge: true));
+  }
+}
